Add files via upload
This is what I have worked on for a bit. I changed the nav bar to look a little cleaner. I created a previous events page. The poker club currently hasn't had any events except for one but they can add to it as they come along. For the forms I added a couple other criteria i believe to be important for understanding users.